backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
chromadb==0.5.15
sentence-transformers==3.2.1
groq==0.11.0
gTTS==2.5.4
pypdf==5.0.1
python-docx==1.1.2
